Move experience files into one folder
