+++ src/App.js
@@ -3,11 +3,11 @@ import { Navigation } from "./pages/navigation/Navigation";
 import { SubNavbar } from "./pages/navigation/SubNavbar";
 import { Home } from "./pages/Home";
 import { About } from "./pages/About";
-import { Experience } from "./pages/Experience";
-import { Work } from "./pages/Work";
-import { Volunteer } from "./pages/Volunteer";
-import { Projects } from "./pages/Projects";
-import { Awards } from "./pages/Awards";
+import { Experience } from "./pages/experience/Experience";
+import { Work } from "./pages/experience/Work";
+import { Volunteer } from "./pages/experience/Volunteer";
+import { Projects } from "./pages/experience/Projects";
+import { Awards } from "./pages/experience/Awards";
 import { Hobbies } from "./pages/hobbies/Hobbies";
 import { Lego } from "./pages/hobbies/Lego";
 import { Art } from "./pages/hobbies/Art";
